I was unhappy with the way that the Map object conflated the idea of the Map itself with a surface for drawing it on. I was also unhappy with the interdependencies between it and the Raycaster, that the Raycaster used it for doing the actual rendering rather than the main display.
To fix this I've refactored Map into two classes, LevelMap which exclusively deals with map-related info, and LevelMapSurface which links a LevelMap to a surface and is only used in Dev Mode.

I've also changed the Raycaster to take the LevelMap on init, and work directly with the main display, rather than drawing to a surface and then handing that back to core to be blit to screen.