Extensive refactor: originally I envisaged GameState as a passive data class which kept track of things, but this resulted in circular imports between Player/GameObject/Level and various subclasses as all of them needed to know the locations of the others for various reasons such as working out whether to attack or not.
This circularity was becoming annoying to work with and the typechecking really highlighted what a bad design it was.

To resolve this I've implemented a much stricter hierarchy. The GameState now deals with triggering behaviours on enemies and bullets, and feeds the behaviour classes the required information. All sprites and the player model are now passive data classes with a few transform methods on them, which are acted upon by behaviours (triggered by the Gamestate) or in the case of the player, the GameState itself (pending further refactor).